+++ android/app/build.gradle.kts
@@ -22,7 +22,7 @@ if (envFile.exists()) {
 }
 
 android {
-    namespace = "com.bp.bread_place"
+    namespace = "com.bp.breadplace"
     compileSdk = flutter.compileSdkVersion
     ndkVersion = "27.0.12077973"
 
@@ -38,7 +38,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId = "com.bp.bread_place"
+        applicationId = "com.bp.breadplace"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = 23
