+++ android/app/src/main/kotlin/com/bp/breadplace/geofence/GeofenceLocationModel.kt
@@ -1,4 +1,4 @@
-package com.bp.bread_place.geofence
+package com.bp.breadplace.geofence
 
 data class GeofenceLocationModel(
     val placeId: String,
