+++ android/app/src/main/kotlin/com/bp/breadplace/geofence/NotificationHelper.kt
@@ -1,4 +1,4 @@
-package com.bp.bread_place.geofence
+package com.bp.breadplace.geofence
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
